+++ DependencyInjection/MassiveSearchExtension.php
@@ -43,6 +43,7 @@ class MassiveSearchExtension extends Extension
         $container->setAlias('massive_search.adapter', $config['adapter_id']);
         $container->setParameter('massive_search.adapter.zend_lucene.basepath', $config['adapters']['zend_lucene']['basepath']);
         $container->setParameter('massive_search.adapter.zend_lucene.hide_index_exception', $config['adapters']['zend_lucene']['hide_index_exception']);
+        $container->setParameter('massive_search.zend_lucene.analyzer_alias', $config['adapters']['zend_lucene']['analyzer']);
 
         $loader->load('search.xml');
     }