+++ MassiveSearchBundle.php
@@ -13,6 +13,7 @@ namespace Massive\Bundle\SearchBundle;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 use Massive\Bundle\SearchBundle\DependencyInjection\Compiler\MetadataDriverPass;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Massive\Bundle\SearchBundle\DependencyInjection\Compiler\ZendLucenePass;
 
 class MassiveSearchBundle extends Bundle
 {
@@ -20,5 +21,6 @@ class MassiveSearchBundle extends Bundle
     {
         parent::build($container);
         $container->addCompilerPass(new MetadataDriverPass());
+        $container->addCompilerPass(new ZendLucenePass());
     }
 }